feat: precio_docena + cantidades fraccionarias (paridad con app v6)
- productos: columna precio_docena (centavos, 0 = no aplica) y stock /
  stock_minimo como decimal(10,2) — la app ya vende media libra (0.5).
- detalle_ventas: cantidad decimal(10,2).
- Producto: fillable + cast integer precio_docena; stock/stock_minimo
  cast float (decimal:2 serializaria string y rompe el cliente RN).
- DetalleVenta: cantidad cast float.
- SyncController: el upsert de productos ya no descarta precio_docena.

Verificado: php -l sin errores en los 5 archivos.

+++ app/Models/DetalleVenta.php
@@ -20,7 +20,8 @@ class DetalleVenta extends Model
     protected function casts(): array
     {
         return [
-            'cantidad' => 'integer',
+            // float: cantidades fraccionarias (media libra = 0.5)
+            'cantidad' => 'float',
             'precio_unitario' => 'integer',
             'subtotal' => 'integer',
         ];
